requirements.txt
Flask==1.1.2
requests==2.24.0
python-dotenv==0.13.0
flask-cors==1.10.3
attrs==19.3.0
